SeREGen/pyproject.toml
[build-system]
requires = ["setuptools>=61.0",]
build-backend = "setuptools.build_meta"

[project]
name = "SeREGen"
version = "0.1.0"
authors = [
  { name="Nikhil Kalidasu", email="srikarnikhil.kalidasu@gmail.com" }
]
description = "Automated, cross-platform Representative Embedding generation."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: GNU General Public License (GPL)",
  "Operating System :: OS Independent"
]
dependencies = [
    "numpy>=1.21.2",
    "tqdm>=4.64.2",
    "torch>=2.3.1",
    "torchtext>=0.17.0",
    "torchinfo>=1.8.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.4.3",
    "biopython>=1.79",
    "pandas>=1.3.3",
    "scipy>=1.10.1",
    "rdkit>=2023.3.3",
    "Levenshtein>=0.25.1",
    "geomstats>=2.7.0",
]

[project.urls]
Homepage = "https://github.com/nik875/SeREGen"
Issues = "https://github.com/nik875/SeREGen/issues"
